created seperate controller for the login and moved login controller implementation from registration controller to login controller
